# config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Masukkan token dengan format: Bearer {token}",
    "description": "Ketika memasukkan token tambahkan Bearer di depannya seperti format diatas",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Masukkan token dengan format: Bearer {token}"
    }
  },
  "security": [
    {
      "bearerToken": []
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "description": "Mendaftarkan pengguna baru",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "johndoe"
                },
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                }
              },
              "required": ["username", "email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "description": "Login pengguna dan mendapatkan token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login berhasil",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                },
                "user": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/profile/update": {
      "put": {
        "tags": ["Profile"],
        "description": "Update profil pengguna",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "phone": {
                  "type": "string",
                  "example": "08123456789"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/profile/image": {
      "put": {
        "tags": ["Profile"],
        "description": "Update foto profil",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "description": "File gambar profil"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile image updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": ["Profile"],
        "description": "Mendapatkan data profil pengguna",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/topup": {
      "post": {
        "tags": ["Transactions"],
        "description": "Top up saldo",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "amount": {
                  "type": "number",
                  "example": 100000
                }
              },
              "required": ["amount"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top up successful"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/balance": {
      "get": {
        "tags": ["Transactions"],
        "description": "Melihat saldo",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "balance": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/transaction/history": {
      "get": {
        "tags": ["Transactions"],
        "description": "Melihat riwayat transaksi",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "date": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/banner": {
      "get": {
        "tags": ["Public"],
        "description": "Mendapatkan daftar banner",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/services": {
      "get": {
        "tags": ["Public"],
        "description": "Mendapatkan daftar layanan",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}